refactor: 식단 swiper 수정
